Create ReviewProgress component and use in start and Review page the same way
