+++ src/components/Review/index.ts
@@ -1,4 +1,5 @@
 import DecisionArea from './DecisionArea'
+import ReviewProgress from './Progress'
 import ReviewSection from './ReviewSection'
 
-export { DecisionArea, ReviewSection }
+export { DecisionArea, ReviewProgress, ReviewSection }
